Created and implemented DeviceBuilder. Renamed DeviceFactory to SensorFactory
